Fix a warning that could occur
When spit or a similar function was used inside a case, cond, or similar
special form, elixir would emit a warning.

This was the case since a name is being rebound inside a different
"scope" (this is not really true, as elixir does not open a new scope
for case, cond, ...) is considered dangerous in elixir.

This was fixed by marking all quoted code with `generated: true`

A unit test with no assertions was added so that it will be immediately
obvious if this solution no longer works in the future