Update component callbacks argument handling
Most of the different callbacks in the DSL had slightly different
conventions for handling arguments. Yet the DSL treated them all the
same. This was not ideal for callbacks such as `restore`, which
generally only accept a single argument, since the user would always
need to pass its argument inside a list. Furthermore, callbacks like
`terminate`, which don't accept any arguments would need a superfluous
empty list in its argument list.

To solve this, the component callback transformation code was reformed,
three compile time lists specify the various categories a callback can
be a part of and the `transform_component_callbacks` function handles
the logic.