Fix an issue with writing to structs in react
The logic for generating the skitter_instance variable in the body of
react could not correctly handle writing to a struct when it was never
read.

To handle this, the strategy for implementing struct updates was
modified slightly, instead of using `instance!/2` a different function:
`instance_field!/2` was created. This makes it possible to correctly
generate the react body.

Besides this, some small changes were needed to accomoddate for this new
logic.