Master / worker / nodes changes
This commit gathers a set of updates made to the nodes, master and
worker modules which were created while working on the component update.
While the nodes module is not complete yet, these changes shoudl form
the basis of the overall layout of the skitter supervision tree.

From now on:

- The master and worker modules define their own supervisers, which
spawn all the required sub-supervisors
- A new local application mode spawn both the master and worker
supervisor, which should make it possible to test skitter workflows on a
single erlang vm
- the nodes module gather all the functionality related to nodes, right
now, the following is available:
  - A registry which tracks the nodes + some meta information
  - A monitor that can send messages when nodes crash
  - A load balancer (which just picks a random node for now) which
  allows you to select a node
  - A task module that makes it easy to execute supervised tasks on
  worker nodes.