Fix import issues in component DSL
Finally figured out the cause of the import weirdness: the import in the
body of defcallback (which does not include defcallback) overrides the
import of defcallback which hapenned inside the component DSL.

This is unavoidable: fix it by always importing defcallback inside the
body of defcallback; this is needed to get handlers to work.
Clean up the code of the component DSL that deals with imports, it is
way less finnicky about scoping rules now.

Also update imports of handler dsl.